backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
pydantic==1.10.14
python-multipart==0.0.9
PyJWT==2.8.0
